requirements.txt
aiofiles==22.1.0
aiosqlite==0.19.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==23.1.0
autopep8==2.0.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
cryptography==40.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.6
django-allauth==0.54.0
django-ckeditor==6.5.1
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
facebook-sdk==3.1.0
fastjsonschema==2.16.3
filelock==3.11.0
fqdn==1.5.1
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.5
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
playsound==1.3.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-openid==2.2.5
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.1
traitlets==5.9.0
tzdata==2022.7
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
y-py==0.5.9
ypy-websocket==0.8.2
